Title: 10개 입력받기, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 231113/10개 입력받기/receive-10-inputs.java
@@ -14,12 +14,11 @@ public class Main {
                 break;
             }else{
                 sum+=n[i];
-                avg=sum/cnt;
+                avg=(double)sum/cnt;
             }
        
         }
             //출력할위치가 맞는지 신경써서 위치시키기
-             System.out.print(String.format("%d %.1f",sum,avg));
-             
+             System.out.print(String.format("%d %.1f",sum,avg));      
     }
 }
